Global and local search cleaned up
