+++ utils/processor.py
@@ -46,90 +46,8 @@ def evaluate_BraggNN(model, train_loader, val_loader, device, num_epochs=50, lr=
     )
     return mean_distance, inference_time, validation_loss, param_count
 
-def evaluate_braggnn(model, train_loader, val_loader, test_loader, device, num_epochs=50, lr=0.0015, weight_decay=2.2e-9):
-    """Evaluates BraggNN models by training and computing performance metrics"""
-    # Move model to device
-    model = model.to(device)
-    
-    # Initialize training components
-    criterion = nn.MSELoss()
-    optimizer = torch.optim.RMSprop(model.parameters(), lr=lr, weight_decay=weight_decay)
-    scheduler = torch.optim.lr_scheduler.CosineAnnealingLR(optimizer, T_max=num_epochs)
-    
-    # Train model
-    validation_loss = train(model, optimizer, scheduler, criterion, 
-                          train_loader, val_loader, device, 
-                          num_epochs)
-    
-    # Calculate metrics
-    val_mean_dist = get_mean_dist(model, val_loader, device, psz=11)
-    test_mean_dist = get_mean_dist(model, test_loader, device, psz=11)
-    param_count = get_param_count_BraggNN(model) # Just for reference, we are not optimizing for this. We measure BOPs in global_search.py
-    inference_time = get_inference_time(model, device, img_size=(256, 1, 11, 11)) # Just for reference, we are not optimizing for this.
-    
-    # Print results
-    # print(
-    #     f"Validation Mean Distance: {val_mean_dist:.4f}, "
-    #     f"Test Mean Distance: {test_mean_dist:.4f}, "
-    #     f"Inference time: {inference_time:.4f}, "
-    #     f"Validation Loss: {validation_loss:.4f}, "
-    #     f"Parameter Count: {param_count}"
-    # )
-    
-    # Return metrics dictionary
-    return {
-        "val_mean_dist": val_mean_dist,
-        "test_mean_dist": test_mean_dist,
-        "val_loss": validation_loss,
-        "inference_time": inference_time,
-        "param_count": param_count
-    }
 
 
-# Trains model and calculates all metrics
-# def evaluate_Deepsets(model, train_loader, val_loader, device, num_epochs=80, lr=0.0032):
-#     model = model.to(device)
-
-#     # Train Model
-#     criterion = nn.CrossEntropyLoss()
-#     optimizer = torch.optim.Adam(model.parameters(), lr=lr)
-#     scheduler = ReduceLROnPlateau(optimizer, mode="min", factor=0.1, patience=3)
-#     validation_loss = train(model, optimizer, scheduler, criterion, train_loader, val_loader, device, num_epochs, patience=7)
-
-#     # Evaluate Performance
-#     acc = get_acc(model, val_loader, device)
-
-#     # Evaluate Efficiency
-#     param_count = get_param_count_Deepsets(model)
-#     inference_time = get_inference_time(
-#         model, device, img_size=(1024, 3, 8)
-#     )  # Just for reference, we are not optimizing for this.
-
-#     print(
-#         "Accuracy: ",
-#         acc,
-#         ", Inference time: ",
-#         inference_time,
-#         ", Validation Loss: ",
-#         validation_loss,
-#         ", Param Count: ",
-#         param_count,
-#     )
-#     return acc, inference_time, validation_loss, param_count
-
-# def evaluate_deepsets_local(model, train_loader, val_loader, test_loader, device):
-#     """Evaluatio function for DeepSets models"""
-#     val_accuracy, inference_time, validation_loss, param_count = evaluate_Deepsets(
-#         model, train_loader, val_loader, device, num_epochs=100
-#     )
-#     test_accuracy = get_acc(model, test_loader, device)
-#     return {
-#         "val_accuracy": val_accuracy,
-#         "test_accuracy": test_accuracy,
-#         "val_loss": validation_loss,
-#         "inference_time": inference_time,
-#         "param_count": param_count
-#     }
 
 def evaluate_deepsets(model, train_loader, val_loader, test_loader, device, num_epochs=100, lr=0.0032):
     """Evaluates DeepSets models by training and computing performance metrics"""
